requirements.txt
appdirs==1.4.4
attrs==21.2.0
click==8.0.1
distlib==0.3.2
filelock==3.0.12
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
inflect==5.3.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
json-schema-to-class==0.2.4
lazy-write==0.0.1
MarkupSafe==2.0.1
mysqlclient==2.0.3
packaging==20.9
pluggy==0.13.1
py==1.10.0
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
six==1.16.0
sqlacodegen==2.3.0
SQLAlchemy==1.4.19
toml==0.10.2
Werkzeug==2.0.1
